Show queue "health" in daemon responses
Also update the client to return the daemon's payload (when status=ok).
Queue percent full will start flowing back to clients as the result of
their log_event() calls.

Also includes a very crude "health status" of "green|yellow|red" at the
0-50%, 50%-70%, 70%-100% boundaries for queue full.

Only examines the frontend queue, as the backend queue is not easily
visible. Furthermore, the backend queue is not filled if its past flush
is not complete, so there is no issue if the cwlogs API calls are all
failing.

Resolves #26