Simplify daemon queue health measurement
- switch from 0.x for percentage to 0.x * 100
- remove the "color statuses" since queue % full is simple on its own
- add _health_info() to the heartbeat events, allowing queue fullness to
  be a LogMetric

We discussed the pros and cons of using a LogMetric for this -- e.g. it
will result in missing data (rather than visible steady growth) if the
logging is failing. However, it will simply and easily expose the
information for metrics, and clients are still free to push a metric
event elsewhere based on the data returned.
Therefore, including both is the easiest option for the daemon.