Use "nr_found" for queue fullness on heartbeat
Computing the queue fullness as `len(_g_queue)` on flush results in
artificially low values for the queue length. Because this only runs
_immediately_ after the queue is emptied, it will almost never appear to
have more than a couple of records. This is despite the fact that the
queue may have been 100% full at the time that it was emptied!

Instead, take the `nr_found` (prior len of _g_queue) and pass it through
to the health_info computation. Requires a minor tweak to let this value
pass through.