Add CONTRIBUTING doc, CODEOWNERS, tweak README
closes #22

- try out a CODEOWNERS file to auto-assign PRs for review

This can also be used to do enforcement, but we should try it a bit (and
get it working!) before we try out those features.

- Add a contributing doc

Covers linting and code standards in short order, and documents the
release process.

- Update the README

Part of the point of CODEOWNERS is to declare maintainers, so we can
drop that line from the readme.

While in there, I also wanted to clarify what this is and what our
posture is towards 3rd party use. I borrowed language from BoringSSL's
readme for this, which has a similar "internal, but open source" mode of
use.