Update readme install commands with egg spec
This is an attempt to close #33 after some testing.
It's hard to handle well without knowing specifically what went wrong
with these installs under a requirements.txt -- taking a guess that one
of the mentioned, commonly-seen issues was at play.

Non-editable install with `-e` or `#` instead of `&`.

A really old version of pip could always break things badly/weirdly, but
that's also a guess.